% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A4Benchmark}
\alias{A4Benchmark}
\title{A4Benchmark}
\format{
A list of time series.
}
\source{
\href{https://doi.org/10.1371/journal.pone.0262463}{Yahoo dataset, 2021}
}
\usage{
data(A4Benchmark)
}
\description{
S5 - A Labeled Anomaly Detection Dataset.
Data Type:	Benchmark A3. Category:	Synthetic data with anomalies and change points. Creation Date	2021.
See \href{https://webscope.sandbox.yahoo.com/catalog.php?datatype=s&did=70}{cefet-rj-dal/dalevents}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\examples{
data(A4Benchmark)
serie <- A4Benchmark[[1]]
}
\keyword{#'}
\keyword{(2022)}
\keyword{17(1).}
\keyword{@references}
\keyword{A}
\keyword{An}
\keyword{K}
\keyword{K,}
\keyword{ONE}
\keyword{PLoS}
\keyword{Takahashi}
\keyword{Web}
\keyword{Yoshihara}
\keyword{anomaly}
\keyword{application}
\keyword{data.}
\keyword{datasets}
\keyword{detection:}
\keyword{for}
\keyword{method}
\keyword{series}
\keyword{simple}
\keyword{time}
\keyword{to}
\keyword{unsupervised}
