% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nab_artificialWithAnomaly}
\alias{nab_artificialWithAnomaly}
\title{Numenta Anomaly Benchmark (NAB) artificialWithAnomaly}
\format{
A list of time series.
}
\source{
\href{https://github.com/numenta/NAB/tree/master/data}{Numenta Anomaly Benchmark (NAB) Dataset}
}
\usage{
data(nab_artificialWithAnomaly)
}
\description{
Data collection with real-world time-series.
Artificial data with anomalies
As part of the Numenta Anomaly Benchmark (NAB), this dataset contains
time series with real and synthetic data. The real data comes from network
monitoring and cloud computing. On the other hand, synthetic data simulate
series with or without anomalies.
See \href{https://github.com/cefet-rj-dal/dalevents}{cefet-rj-dal/dalevents}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\examples{
data(nab_artificialWithAnomaly)
data <- nab_artificialWithAnomaly[[1]]
series <- data$value
}
\keyword{#'#'}
\keyword{@references}
\keyword{Numenta}
\keyword{datasets}
