% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ucr_nasa}
\alias{ucr_nasa}
\title{UCR Anomaly Benchmark - NASA spacecraft series}
\format{
A list of time series.
}
\source{
\href{https://paperswithcode.com/dataset/ucr-anomaly-archive/}{UCI Anomaly Archive}
}
\usage{
data(ucr_nasa)
}
\description{
Data collection with real-world time-series.
Real data from NASA Spacecraft Monitoring Data
See \href{https://github.com/cefet-rj-dal/dalevents}{cefet-rj-dal/dalevents}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\examples{
data(ucr_nasa)
data <- ucr_nasa[[1]]
series <- data$value
}
\keyword{#'#'}
\keyword{@references}
\keyword{Anomaly}
\keyword{Archive}
\keyword{UCR}
\keyword{datasets}
