% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ucr_power_demand}
\alias{ucr_power_demand}
\title{UCR Anomaly Benchmark - Italian Power Demand}
\format{
A list of time series.
}
\source{
\href{https://paperswithcode.com/dataset/ucr-anomaly-archive/}{UCI Anomaly Archive}
}
\usage{
data(ucr_power_demand)
}
\description{
Data collection with real-world time-series.
Real data from Italian Power Demand
See \href{https://github.com/cefet-rj-dal/dalevents}{cefet-rj-dal/dalevents}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\examples{
data(ucr_power_demand)
data <- ucr_power_demand[[1]]
series <- data$value
}
\keyword{#'#'}
\keyword{@references}
\keyword{Anomaly}
\keyword{Archive}
\keyword{UCR}
\keyword{datasets}
